Delte unused System.ComponentModel.Composition code in favor of shared code
