+++ src/System.ComponentModel.Composition/src/System.ComponentModel.Composition.csproj
@@ -82,11 +82,7 @@
     <Compile Include="System\ComponentModel\Composition\AttributedModel\AttributedExportDefinition.cs" />
     <Compile Include="System\ComponentModel\Composition\AttributedModel\AttributedModelDiscovery.cs" />
     <Compile Include="System\ComponentModel\Composition\AttributedModel\AttributedPartCreationInfo.cs" />
-    <Compile Include="System\ComponentModel\Composition\Diagnostics\CompositionTrace.cs" />
-    <Compile Include="System\ComponentModel\Composition\Diagnostics\CompositionTraceId.cs" />
-    <Compile Include="System\ComponentModel\Composition\Diagnostics\CompositionTraceSource.cs" />
-    <Compile Include="System\ComponentModel\Composition\Diagnostics\DebuggerTraceWriter.cs" />
-    <Compile Include="System\ComponentModel\Composition\Diagnostics\TraceWriter.cs" />
+    <Compile Include="System\Composition\Diagnostics\CompositionTrace.cs" />
     <Compile Include="System\ComponentModel\Composition\Hosting\AggregateCatalog.cs" />
     <Compile Include="System\ComponentModel\Composition\Hosting\AggregateExportProvider.cs" />
     <Compile Include="System\ComponentModel\Composition\Hosting\ApplicationCatalog.cs" />
@@ -184,6 +180,18 @@
     <Compile Include="System\ComponentModel\Composition\ReflectionModel\ReflectionProperty.cs" />
     <Compile Include="System\ComponentModel\Composition\ReflectionModel\ReflectionType.cs" />
     <Compile Include="System\ComponentModel\Composition\ReflectionModel\ReflectionWritableMember.cs" />
+    <Compile Include="$(CommonPath)\System\Composition\Diagnostics\CompositionTraceId.cs">
+      <Link>Common\System\Composition\Diagnostics\CompositionTraceId.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\System\Composition\Diagnostics\CompositionTraceSource.cs">
+      <Link>Common\System\Composition\Diagnostics\CompositionTraceSource.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\System\Composition\Diagnostics\DebuggerTraceWriter.cs">
+      <Link>Common\System\Composition\Diagnostics\DebuggerTraceWriter.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\System\Composition\Diagnostics\TraceWriter.cs">
+      <Link>Common\System\Composition\Diagnostics\TraceWriter.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' != 'netstandard'">
     <Reference Include="Microsoft.Win32.Primitives" />
