--- src/System.ComponentModel.Composition/src/System/ComponentModel/Composition/Diagnostics/CompositionTraceSource.cs
@@ -1,56 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-
-namespace System.ComponentModel.Composition.Diagnostics
-{
-    internal static class CompositionTraceSource
-    {
-        private static readonly DebuggerTraceWriter Source = new DebuggerTraceWriter();
-
-        public static bool CanWriteInformation
-        {
-            get { return Source.CanWriteInformation; }
-        }
-
-        public static bool CanWriteWarning
-        {
-            get { return Source.CanWriteWarning; }
-        }
-
-        public static bool CanWriteError
-        {
-            get { return Source.CanWriteError; }
-        }
-
-        public static void WriteInformation(CompositionTraceId traceId, string format, params object[] arguments)
-        {
-            EnsureEnabled(CanWriteInformation);
-
-            Source.WriteInformation(traceId, format, arguments);
-        }
-
-        public static void WriteWarning(CompositionTraceId traceId, string format, params object[] arguments)
-        {
-            EnsureEnabled(CanWriteWarning);
-
-            Source.WriteWarning(traceId, format, arguments);
-        }
-
-        public static void WriteError(CompositionTraceId traceId, string format, params object[] arguments)
-        {
-            EnsureEnabled(CanWriteError);
-
-            Source.WriteError(traceId, format, arguments);
-        }
-
-        private static void EnsureEnabled(bool condition)
-        {
-            if(!condition)
-            {
-                throw new Exception(SR.Trace_Level_Not_Enabled);
-            }
-        }
-    }
-}
