+++ src/System.ComponentModel.Composition/src/System/Composition/Diagnostics/CompositionTrace.cs
@@ -2,13 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.ComponentModel.Composition;
 using System.ComponentModel.Composition.Hosting;
 using System.ComponentModel.Composition.Primitives;
 using System.ComponentModel.Composition.ReflectionModel;
 using System.Reflection;
 using Microsoft.Internal;
 
-namespace System.ComponentModel.Composition.Diagnostics
+namespace System.Composition.Diagnostics
 {
     internal static class CompositionTrace
     {
